Comment Repository 수정
Comment Repository 수정

+++ metanet_internship_4team/src/main/java/com/metanet/repository/CommentsRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.metanet.domain.Comments;
 
 @Repository
-public interface CommentsRepository extends JpaRepository<Comments, Long>
+public interface CommentsRepository extends JpaRepository<Comments, Integer>
 
 {
 	List<Comments> findByvideoNumber(int videoNumber);
